requirements.txt
streamlit==1.32.0
pydantic>=2.6.0
langchain==0.0.335
langchain-google-genai==0.0.7
langgraph>=0.0.19
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
duckduckgo-search==3.9.9
google-generativeai==0.3.2
watchdog>=3.0.0
nltk>=3.8.1
